Rename invoiceninja -> invoice_ninja
